Filter aggregated expenditures by election.
