+++ tests/test_utils.py
@@ -6,7 +6,6 @@ from webargs import flaskparser
 from tests import factories
 from tests.common import ApiBaseTest
 
-from webservices import utils
 from webservices import args
 from webservices import rest
 from webservices import sorting
@@ -62,48 +61,6 @@ class TestSort(ApiBaseTest):
         self.assertEqual(query.all(), candidates[-1:] + candidates[:2])
         query, columns = sorting.sort(models.Candidate.query, '-district', model=models.Candidate, nulls_large=False)
         self.assertEqual(query.all(), [candidates[1], candidates[0], candidates[2]])
-
-
-class TestFilter(ApiBaseTest):
-
-    def setUp(self):
-        super(TestFilter, self).setUp()
-        self.receipts = [
-            factories.ScheduleAFactory(entity_type='IND'),
-            factories.ScheduleAFactory(entity_type='CCM'),
-            factories.ScheduleAFactory(entity_type='COM'),
-            factories.ScheduleAFactory(entity_type='PAC'),
-        ]
-
-    def test_filter_contributor_type_individual(self):
-        query = utils.filter_contributor_type(
-            models.ScheduleA.query,
-            models.ScheduleA.entity_type,
-            {'contributor_type': ['individual']},
-        )
-        self.assertEqual(
-            set(query.all()),
-            set(each for each in self.receipts if each.entity_type == 'IND')
-        )
-
-    def test_filter_contributor_type_committee(self):
-        query = utils.filter_contributor_type(
-            models.ScheduleA.query,
-            models.ScheduleA.entity_type,
-            {'contributor_type': ['committee']},
-        )
-        self.assertEqual(
-            set(query.all()),
-            set(each for each in self.receipts if each.entity_type != 'IND')
-        )
-
-    def test_filter_contributor_type_none(self):
-        query = utils.filter_contributor_type(
-            models.ScheduleA.query,
-            models.ScheduleA.entity_type,
-            {'contributor_type': ['individual', 'committee']},
-        )
-        self.assertEqual(set(query.all()), set(self.receipts))
 
 
 class TestArgs(unittest.TestCase):
